+++ filament/backend/include/backend/platforms/PlatformEGLAndroid.h
@@ -86,6 +86,7 @@ protected:
     void attach(Stream* stream, intptr_t tname) noexcept override;
     void detach(Stream* stream) noexcept override;
     void updateTexImage(Stream* stream, int64_t* timestamp) noexcept override;
+    math::mat3f getTransformMatrix(Stream* stream) noexcept override;
 
     /**
      * Converts a AHardwareBuffer to EGLImage