+++ filament/src/details/MaterialInstance.cpp
@@ -24,6 +24,7 @@
 #include "details/Material.h"
 #include "details/MaterialInstance.h"
 #include "details/Texture.h"
+#include "details/Stream.h"
 
 #include "private/filament/EngineEnums.h"
 
@@ -190,9 +191,27 @@ void FMaterialInstance::terminate(FEngine& engine) {
     driver.destroyBufferObject(mUbHandle);
 }
 
+void FMaterialInstance::commitStreamUniformAssociations(FEngine::DriverApi& driver) {
+    mHasStreamUniformAssociations = false;
+    if (!mTextureParameters.empty()) {
+        backend::BufferObjectStreamDescriptor descriptor;
+        for (auto const& [binding, p]: mTextureParameters) {
+            ssize_t offset = mMaterial->getUniformInterfaceBlock().getTransformFieldOffset(binding);
+            if (offset >= 0) {
+                mHasStreamUniformAssociations = true;
+                auto stream = p.texture->getStream()->getHandle();
+                descriptor.mStreams.push_back({uint32_t(offset), stream, BufferObjectStreamAssociationType::TRANSFORM_MATRIX});
+            }
+        }
+        if (descriptor.mStreams.size() > 0) {
+            driver.registerBufferObjectStreams(mUbHandle, std::move(descriptor));
+        }
+    }
+}
+
 void FMaterialInstance::commit(DriverApi& driver) const {
     // update uniforms if needed
-    if (mUniforms.isDirty()) {
+    if (mUniforms.isDirty() || mHasStreamUniformAssociations) {
         driver.updateBufferObject(mUbHandle, mUniforms.toBufferDescriptor(driver), 0);
     }
     if (!mTextureParameters.empty()) {