+++ tools/matinfo/src/main.cpp
@@ -423,7 +423,7 @@ static bool parseChunks(Config config, void* data, size_t size) {
     if (config.serverPort) {
         // Spin up a web server on a secondary thread.
         DebugServer server(Backend::DEFAULT, filament::backend::ShaderLanguage::ESSL3,
-                config.serverPort);
+                filament::matdbg::DbgShaderModel::MATINFO, config.serverPort);
         if (!server.isReady()) {
             return false;
         }
@@ -532,7 +532,7 @@ static bool parseChunks(Config config, void* data, size_t size) {
 
             return true;
         }
-
+        //TODO Include a printWgsl logic here
         if (config.printMetal) {
             const filament::backend::ShaderLanguage language = config.binary
                     ? filament::backend::ShaderLanguage::METAL_LIBRARY